.gitignore file fixed duration to seconds fixed most watches ordering to descending

+++ Moviesapi/MovieStats/GetMovieStats/MovieStatResopnseModel.cs
@@ -4,7 +4,7 @@
     {
         public int MovieId { get; set; }
         public string Title { get; set; }
-        public float AverageWatchDurationS { get; set; }
+        public int AverageWatchDurationS { get; set; }
         public int Watches { get; set; }
         public int ReleaseYear { get; set; }
     }
